Same events at the same time don't replace each other
Previously, if multiple events of the same type were scheduled at the
same time, then each new event would replace the existing event.  This
is incorrect according to the spec.  There is no special treatment
here and events are inserted in the order in which they're received.

This change only affects linear and exponential ramps.  The result is
that the ramp preceeds up to the first event value, and then at the
event time, the output instantly jumps to the last event value.

Bug: 925037
Test: the-audioparam-interface/event-insertion.html updated
Change-Id: I2de32c7a71c12e5673b2db8cf81f9dc48f3ee458