[Native File System] Change at what level we hook into FS backend.
Before this CL the Native File System API hooked into the old FS api
from within blink. This makes implementing several of the features we want
to support in the new API very difficult, so this CL changes things so
the integration point now is in the browser process.

As part of doing this, code specific to the new Native File System API is
also moved to its own directories in content and blink.

Also get rid of getParent() and update tests to not rely on getParent()
as that method is not part of the planned API surface.

Bug: 853326
Binary-Size: Increase is due to large size of mojo bindings needed to
    implement a new feature.

Change-Id: Ia2702f83b6c80bb4f2d014b48e251814677a59bc