Portals: Implement PortalHost.postMessage
Implements a version of postMessage for PortalHost that can can only
send strings. It implements messaging from the portal to the embedding
page, i.e. in the opposite direction to https://crrev.com/c/1423462.
Support for more general message types will be added in a follow up CL.

This CL introduces 2 new mojo interfaces:

blink.mojom.PortalClient: Allows content::Portal in the browser process
to send messages to the corresponding HTMLPortalElement in the outer
renderer process. (This is in the opposite direction as
blink.mojom.Portal, which goes from renderer to browser.)

blink.mojom.PortalHost: Used by the blink::PortalHost object (in the
inner renderer process) to send messages to the content::Portal in the
browser process.

Both the above interfaces together allow blink::PortalHost in the inner
renderer process to send messages to the blink::HTMLPortalElement in
the outer renderer process.

Bug: 914120
Change-Id: If7dbaf6e2817fcc94dc38a1df25afc25d8b75eb3