[IntersectionObserver] Correctly handle a scaled target
If a target is inside an iframe, and the iframe (or one if its
ancestors in the parent document) has a scale applied, then we need to
apply the scale to the hit test rect when checking for occlusion. The
default LocalFrameView coordinate conversion routines don't do that.
Use LocalToAncestorQuad instead.

BUG=951525
R=chrishtr@chromium.org

Change-Id: I1107f0bd3c6d262a392fbafe430897cc09068623